compile infer function based on platform
